Check if CONF exists before trying to remove it
This fixes a case where the container won't start because it can't remove the file.

+++ docker-entrypoint.sh
@@ -3,7 +3,7 @@ set -e
 
 # mail setup
 CONF=/etc/ssmtp/ssmtp.conf
-rm $CONF
+[ -e $CONF ] && rm $CONF
 
 for E in $(env)
 do
